backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
mercadopago